+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -321,7 +321,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KT8Z5A1GSG1VM8MVZXGQRKRN"
+        "01KT94XBKZNX3FJ0W5283YQX9A"
     )
     r"""The unique identifier of the data source"""