+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KT8Z5A1F7AS5WP8K1T469E6C"
+        "01KT94XBKXR4HB0P4A2FDH8RGR"
     )
     r"""The unique identifier of the data source"""